% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_operations.R
\name{age_cal}
\alias{age_cal}
\title{Age Calculator}
\usage{
age_cal(data, variable)
}
\arguments{
\item{data}{name of dataset}

\item{variable}{variable name whose age is to be calculated}
}
\value{
a dataframe with age as a vector
}
\description{
Calculate age variable in years from an existing calendar year variable in dataset by subtracting year variable from System Date
}
\examples{
csv = system.file("extdata", 'agedata.csv', package = "nacleanr")
agedata <- read_data(csv)
agedata$ageTodaySinceBuilt <- age_cal(agedata,"YearBuilt")
agedata$ageTodaySinceRenovated <- age_cal(agedata, "YearRenovated")
head(agedata)
}
