% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_operations.R
\name{invalidcols}
\alias{invalidcols}
\title{Return Invalid Variable Names}
\usage{
invalidcols(data, threshold)
}
\arguments{
\item{data}{name of dataset}

\item{threshold}{threshold percentage of NAs or missing values}
}
\description{
Return Invalid Variable Names
}
