% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read CSV File and convert missing values in form of spaces, empty value, character NA and na into logical NAs}
\usage{
read_data(path)
}
\arguments{
\item{path}{"path/filename.csv"}
}
\value{
Second sample data
}
\description{
Read CSV File and convert missing values in form of spaces, empty value, character NA and na into logical NAs
}
\examples{
csv = system.file("extdata", 'nadata.csv', package = "nacleanr")
read_data(csv)
}
