% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_operations.R
\name{select_cols}
\alias{select_cols}
\title{Valid Variable Selector}
\usage{
select_cols(data, threshold)
}
\arguments{
\item{data}{name of dataset}

\item{threshold}{threshold percentage of NAs or missing values}
}
\value{
output dataset after operation
}
\description{
Valid Variable Selector
}
