% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_cal.R, R/variable_operations.R
\name{age_cal}
\alias{age_cal}
\title{Age Calculator}
\usage{
age_cal(data, variable)

age_cal(data, variable)
}
\arguments{
\item{data}{name of dataset}

\item{variable}{variable name whose age is to be calculated}

\item{agevariable}{new variable column which is the calculated age}

\item{data}{name of dataset}

\item{variable}{variable name whose age is to be calculated}
}
\value{
a dataframe with age as a vector

a dataframe with age as a vector
}
\description{
Age Calculator

Age Calculator
}
