% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_operations.R
\name{validcols}
\alias{validcols}
\title{Return Valid variable names}
\usage{
validcols(data, threshold)
}
\arguments{
\item{data}{name of dataset}

\item{threshold}{threshold percentage of NAs or missing values}
}
\value{
names of variables that fall within user defined criteria for missing value
}
\description{
Return Valid variable names
}
\examples{
csv = system.file("extdata", 'nadata.csv', package = "nacleanr")
data <- read_data(csv)
validcols(data,50)
}
